requirements.txt
# Core Web Framework
fastapi==0.127.1
uvicorn==0.40.0
python-multipart==0.0.21
starlette==0.50.0

# Machine Learning
tensorflow==2.20.0
keras==3.13.0
numpy==2.2.6
scikit-learn==1.8.0

# Image Processing
pillow==12.0.0
opencv-python==4.12.0.88

# Utilities
requests==2.32.5
python-dateutil==2.9.0.post0

# TensorFlow Dependencies
absl-py==2.3.1
astunparse==1.6.3
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
libclang==18.1.1
ml_dtypes==0.5.4
opt_einsum==3.4.0
protobuf==6.33.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
wrapt==2.0.1

# NVIDIA CUDA Support (for potential GPU instances)
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.17.1.4
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.9.86
